% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bridges.R
\name{spectral_bridges}
\alias{spectral_bridges}
\title{Spectral Bridges Clustering}
\usage{
spectral_bridges(X, n_classes = NULL, n_cells = NULL, p = 2, M = 10000)
}
\arguments{
\item{X}{A numeric matrix of data to be clustered.}

\item{n_classes}{Number of classes to cluster into. If NULL, the number of classes will be determined automatically.}

\item{n_cells}{Number of cells of Voronoi tessellation. If NULL, a heuristic is used.}

\item{p}{Power parameter for affinity computation.}

\item{M}{A parameter for the transformation.}
}
\value{
A list containing the clustering result and the original data.
}
\description{
Performs spectral bridges clustering on the given dataset.
}
\examples{
\dontrun{
library(spectralBridges)
X <- iris[, 1:4]
True_classes <- iris$Species
res <- spectral_bridges(X, n_cells = 12, n_classes = 3)
table(True_classes, Est_classes = res$clustering)
}

}
